pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentprobe"
version = "0.1.0"
description = "Multi-agent quality governance system for AI-generated code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "tree-sitter>=0.21.0,<0.22.0",
    "tree-sitter-languages>=1.10.0",
    "PyGithub>=2.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.98.0",
]
experimental = [
    "numpy>=1.26.0",
    "scipy>=1.11.0",
]
experimental-torch = [
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "torch>=2.1.0",
]
experimental-viz = [
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
